# Micro-optimize fuse. `next(iter(myset))` is much slower than a pop then add