Illustrative micro-benchmark:
```python
In [1]: import random

In [2]: L = list(range(100000))

In [3]: random.shuffle(L)

In [4]: s = set(L)

In [5]: %%time
   ...: while s:
   ...:     parent = next(iter(s))
   ...:     s.remove(parent)
   ...:_
CPU times: user 2.32 s, sys: 8 ms, total: 2.32 s
Wall time: 2.32 s

In [6]: s = set(L)

In [7]: %%time
   ...: while s:
   ...:     parent = s.pop()
   ...:     s.add(parent)
   ...:     s.remove(parent)
   ...:_
CPU times: user 32 ms, sys: 0 ns, total: 32 ms
Wall time: 30.3 ms
```